Move register command into its own
